Updated some stuff...

+++ app/src/main/java/com/badlogic/androidgames/framework/Joystick.java
@@ -4,8 +4,8 @@ import java.util.List;
 
 public interface Joystick extends Button {
     List<Input.TouchEvent> processAndRelease(List<Input.TouchEvent> events);
-    double getAngle();
-    double getDistance();
-    double getNormX();
-    double getNormY();
+    float getAngle();
+    float getDistance();
+    float getNormX();
+    float getNormY();
 }
